requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
boto3~=1.26.112
botocore~=1.29.112
certifi==2022.12.7
charset-normalizer==3.1.0
discord-py-interactions==5.6.0
emoji==2.2.0
frozenlist==1.3.3
idna==3.4
jmespath==1.0.1
multidict==6.0.4
openai~=0.27.4
python-dateutil==2.8.2
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
python-dotenv~=1.0.0
peewee~=3.16.2
injector~=0.20.1